Add a cxx_extension_library - a shared library that's statically linked to all of it's dependencies.
The intent is - similar to cxx_python_extension - there are programs that expect to load an .so as a plugin, but we don't want a full shared library build. For example,
MySQL UDFs. This rule builds a shared library with all dependencies statically linked in so, similarly to cxx_binary, you have only 1 build artifact.